Moved functionality to separate files for better readability and testing
